C#: Move nuget related `DependencyManager` methods to separate file
